suppress some warnings on tests

+++ test/src/test_mat2.h
@@ -81,7 +81,7 @@ TEST_IMPL(GLM_PREFIX, mat2_identity) {
 }
 
 TEST_IMPL(GLM_PREFIX, mat2_identity_array) {
-  int i, count;
+  size_t i, count;
   mat2 matrices[4] = {
     A_MATRIX2x2,
     A_MATRIX2x2,
